\name{fastCI-package}
\alias{fastCI-package}
\alias{fastCI}
\docType{package}
\title{
\packageTitle{fastCI}
}
\description{
\packageDescription{fastCI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fastCI}
\packageIndices{fastCI}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fastCI}

Maintainer: \packageMaintainer{fastCI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
